requirements.txt
lit
mypy
numpy
pyyaml
mlir-snapshot
